Generate a Simplified Root Cert in Memory
This commit utilizes some classes including `X509`, `X509Builder`, `X509Name`,
and `X509NameBuilder` to set some preliminary values into a name and a
certificate and generate it. This is not expected to be usable at the
moment. It is more for exploring the available components within the
`OpenSSL` library.